+++ partiql-types/src/main/kotlin/org/partiql/value/impl/PartiQLValueNullableImpl.kt
@@ -317,7 +317,7 @@ internal data class NullableBlobValueImpl(
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
-        other as NullableClobValueImpl
+        other as NullableBlobValueImpl
         if (value != null) {
             if (other.value == null) return false
             if (!value.contentEquals(other.value)) return false